Split out the interrupt handlers into different files
This is cleaner and reduces the need for macros in the codebase.
